Fix bug in mulhw simulation introduced by 36e1509

+++ src/ppc/simulator-ppc.cc
@@ -1916,8 +1916,8 @@ void Simulator::DecodeExt2_9bit(Instruction* instr) {
       int rt = instr->RTValue();
       int ra = instr->RAValue();
       int rb = instr->RBValue();
-      int32_t ra_val = (get_register(ra) & 0xFFFFFFFF);
-      int32_t rb_val = (get_register(rb) & 0xFFFFFFFF);
+      int64_t ra_val = (get_register(ra) & 0xFFFFFFFF);
+      int64_t rb_val = (get_register(rb) & 0xFFFFFFFF);
       int64_t alu_out = ra_val * rb_val;
       alu_out >>= 32;
       set_register(rt, alu_out);
